Int is enough.

+++ src/AdventOfCode/2022/Day_01.cs
@@ -5,9 +5,9 @@ public class Day_01
 {
     [Example(answer: 24000, "1000;2000;3000;;4000;;5000;6000;;7000;8000;9000;;10000")]
     [Puzzle(answer: 69528)]
-    public long part_one(string input) => input.GroupedLines().Select(lines => lines.Int32s().Sum()).Max();
+    public int part_one(string input) => input.GroupedLines().Select(lines => lines.Int32s().Sum()).Max();
 
     [Example(answer: 45000, "1000;2000;3000;;4000;;5000;6000;;7000;8000;9000;;10000")]
     [Puzzle(answer: 206152)]
-    public long part_two(string input) => input.GroupedLines().Select(lines => lines.Int32s().Sum()).OrderByDescending(m => m).Take(3).Sum();
+    public int part_two(string input) => input.GroupedLines().Select(lines => lines.Int32s().Sum()).OrderByDescending(m => m).Take(3).Sum();
 }
